feat: 로그인 상태 시 장바구니 도서 삭제 및 주문 내역 조회가 가능하도록 기능 구현

+++ controllers/cartController.js
@@ -65,6 +65,18 @@ const getCartItems = (req, res) => {
 };
 
 const removeCartItem = (req, res) => {
+  const authorization = ensureAuthorization(req, res);
+
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: '로그인 세션이 만료되었습니다. 다시 로그인 하세요.',
+    });
+  }
+  if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: '잘못된 토큰입니다.',
+    });
+  }
   const cartItemId = req.params.id;
 
   const sql = 'DELETE FROM cartItems WHERE id=?;';
